+++ src/main/kotlin/br/com/gamemods/nbtmanipulator/NbtIO.kt
@@ -13,28 +13,55 @@ import kotlin.reflect.KClass
  */
 object NbtIO {
     /**
-     * Writes the [NbtFile] in the stream.
+     * Calls [writeNbtFile] using the information stored in the [NbtFile], uses the method's default when the information
+     * is missing (null).  This method does not write the Bedrock Edition version and length headers.
      * @param outputStream The stream that the file will be written
-     * @param file The file that will be written on the stream
-     * @param compressed If the file will be compressed by [GZIPOutputStream].
+     * @param file The file that will be written to the stream
      */
     @JvmStatic
     @Throws(IOException::class)
-    fun writeNbtFile(outputStream: OutputStream, file: NbtFile, compressed: Boolean = true) {
-        val tag = file.tag
-        val typeId = tag.typeId
-        val serializer = serializers[typeId]
+    fun writeNbtFileAsOriginal(outputStream: OutputStream, file: NbtFile) {
+        writeNbtFile(outputStream, file, 
+            compressed = file.isCompressed ?: true,
+            littleEndian = file.isLittleEndian ?: false
+        )
+    }
+    
+    /**
+     * Writes the [NbtFile] in the stream. This method does not write the Bedrock Edition version and length headers.
+     * @param outputStream The stream that the file will be written
+     * @param file The file that will be written to the stream
+     * @param compressed If the file will be compressed by [GZIPOutputStream].
+     * @param littleEndian Uses little endian to write to the stream as in Bedrock Edition
+     */
+    @JvmStatic
+    @Throws(IOException::class)
+    @JvmOverloads
+    fun writeNbtFile(outputStream: OutputStream, file: NbtFile, compressed: Boolean = true, littleEndian: Boolean = false) {
         val output = if (compressed) GZIPOutputStream(outputStream) else outputStream
-        val dataOut = DataOutputStream(output)
-
-        dataOut.writeByte(typeId)
-        dataOut.writeUTF(file.name)
-
-        serializer.writeTag(dataOut, tag)
-        dataOut.flush()
+        val dataOut: DataOutput = if (littleEndian) LittleEndianDataOutputStream(output) else DataOutputStream(output)
+        writeNbtFileDirectly(dataOut, file)
+        (dataOut as OutputStream).flush()
         if (output is GZIPOutputStream) {
             output.finish()
         }
+    }
+
+    /**
+     * Writes the [NbtFile] to the output. This method does not write the Bedrock Edition version and length headers.
+     * @param output Where the file will be written, needs to handle compression and endianness.
+     * @param file The file that will be written to the output
+     */
+    @JvmStatic
+    @Throws(IOException::class)
+    fun writeNbtFileDirectly(output: DataOutput, file: NbtFile) {
+        val tag = file.tag
+        val typeId = tag.typeId
+        val serializer = serializers[typeId]
+        output.writeByte(typeId)
+        output.writeUTF(file.name)
+
+        serializer.writeTag(output, tag)
     }
 
     /**
@@ -42,41 +69,153 @@ object NbtIO {
      * @param file The output file
      * @param file The NBT file that will be written on the output file
      * @param compressed If the file will be compressed by [GZIPOutputStream]
+     * @param littleEndian Uses little endian to write to the stream as in Bedrock Edition
+     * @param writeHeaders Writes the [NbtFile.version] and content size headers to the file.
+     * The [NbtFile.length] property will be updated when this flag is set to true.
+     * If [NbtFile.version] is null when this flag is true, `0` is assumed.
+     * The header is always written in little endian regardless of the [littleEndian] param.
      */
     @JvmStatic
     @Throws(IOException::class)
-    fun writeNbtFile(file: File, tag: NbtFile, compressed: Boolean = true) {
-        file.outputStream().buffered().use { writeNbtFile(it, tag, compressed); it.flush() }
+    @JvmOverloads
+    fun writeNbtFile(
+        file: File, tag: NbtFile, compressed: Boolean = true, 
+        littleEndian: Boolean = false, writeHeaders: Boolean = false
+    ) {
+        if (!writeHeaders) {
+            file.outputStream().buffered().use { stream ->
+                writeNbtFile(stream, tag, compressed, littleEndian)
+                stream.flush()
+            }
+        } else {
+            RandomAccessFile(file, "rw").use { openFile ->
+                openFile.setLength(8)
+                FileOutputStream(openFile.fd).buffered().let { stream ->
+                    with(LittleEndianDataOutputStream(stream)) {
+                        writeInt(tag.version ?: 0)
+                        writeInt(0)
+                        flush()
+                    }
+                    writeNbtFile(stream, tag, compressed, littleEndian)
+                    stream.flush()
+                }
+                
+                val fileLength = openFile.length() - 8L
+                val intLength = if (fileLength > Int.MAX_VALUE) Int.MAX_VALUE else fileLength.toInt()
+                tag.length = intLength
+                
+                openFile.seek(8)
+                with(LittleEndianDataOutputStream(FileOutputStream(openFile.fd))) {
+                    writeInt(intLength)
+                    flush()
+                }
+            }
+        }
     }
 
     /**
      * Read a [NbtFile] from the [InputStream].
      * @param inputStream The input stream that will be read
      * @param compressed If the file needs to be decompressed by [GZIPInputStream]
+     * @param littleEndian Reads the NBT file using little endian byte order
+     * @param readHeaders Reads the NBT version and length headers before the content
+     * These data are read in little endian byte order regardless of the [littleEndian] parameter.
      */
     @JvmStatic
     @Throws(IOException::class)
-    fun readNbtFile(inputStream: InputStream, compressed: Boolean = true): NbtFile {
-        val input = if (compressed) GZIPInputStream(inputStream) else inputStream
-        val dataIn = DataInputStream(input)
+    @JvmOverloads
+    fun readNbtFile(
+        inputStream: InputStream, compressed: Boolean = true, 
+        littleEndian: Boolean = false, readHeaders: Boolean = false
+    ): NbtFile {
+        var version: Int? = null
+        var length: Int? = null
 
-        val typeId = dataIn.read()
+        if (readHeaders) {
+            with(LittleEndianDataInputStream(inputStream)) {
+                version = readInt()
+                length = readInt()
+            }
+        }
+
+        val input = if (compressed) GZIPInputStream(inputStream) else inputStream
+        val dataIn: DataInput = if (littleEndian) LittleEndianDataInputStream(input) else DataInputStream(input)
+        val nbtFile = readNbtFileDirectly(dataIn)
+        nbtFile.version = version
+        nbtFile.length = length
+        nbtFile.isCompressed = compressed
+        nbtFile.isLittleEndian = littleEndian
+        return nbtFile
+    }
+
+    /**
+     * Reads a [NbtFile] from the input. This method does not read the Bedrock Edition version and length headers.
+     * @param input Where the file will be read, needs to handle compression and endianness.
+     */
+    @JvmStatic
+    @Throws(IOException::class)
+    fun readNbtFileDirectly(input: DataInput): NbtFile {
+        val typeId = input.readUnsignedByte()
         val serializer = serializers[typeId]
 
-        val name = dataIn.readUTF()
+        val name = input.readUTF()
 
-        return NbtFile(name, serializer.readTag(dataIn))
+        return NbtFile(name, serializer.readTag(input))
     }
 
     /**
      * Read a [NbtFile] from a [File].
      * @param file The input file that will be read
      * @param compressed If the file needs to be decompressed by [GZIPInputStream]
+     * @param littleEndian
+     * @param readHeaders
      */
     @JvmStatic
     @Throws(IOException::class)
-    fun readNbtFile(file: File, compressed: Boolean = true): NbtFile {
-        return file.inputStream().buffered().use { readNbtFile(it, compressed) }
+    @JvmOverloads
+    fun readNbtFile(
+        file: File, compressed: Boolean = true,
+        littleEndian: Boolean = false, 
+        readHeaders: Boolean = false
+    ): NbtFile {
+        return file.inputStream().buffered().use { readNbtFile(it, compressed, littleEndian, readHeaders) }
+    }
+
+    
+    /**
+     * Does an exhaustive attempts to load the NBT file, returning it if any of the attempts is successful. 
+     * @param file
+     */
+    @JvmStatic
+    @Throws(IOException::class)
+    fun readNbtFileDetectingSettings(file: File): NbtFile {
+        RandomAccessFile(file, "r").use { openFile ->
+            var ex: IOException? = null
+            fun retry(compressed: Boolean, littleEndian: Boolean, readHeaders: Boolean): NbtFile? {
+                return try {
+                    openFile.seek(0)
+                    readNbtFile(FileInputStream(openFile.fd).buffered(), compressed, littleEndian, readHeaders)
+                }  catch (e: IOException) {
+                    val lastEx = ex
+                    if (lastEx == null) {
+                        ex = lastEx
+                    } else {
+                        lastEx.addSuppressed(e)
+                    }
+                    null
+                }
+            }
+            return retry(compressed = true, littleEndian = false, readHeaders = false)  // Java's level.dat
+                ?: retry(compressed = false, littleEndian = true, readHeaders = true)   // Bedrock's level.dat
+                ?: retry(compressed = true, littleEndian = false, readHeaders = true)   // Trying all possibilities
+                ?: retry(compressed = true, littleEndian = true, readHeaders = false)
+                ?: retry(compressed = true, littleEndian = true, readHeaders = true)
+                ?: retry(compressed = false, littleEndian = false, readHeaders = false)
+                ?: retry(compressed = false, littleEndian = false, readHeaders = true)
+                ?: retry(compressed = false, littleEndian = true, readHeaders = false)
+                ?: retry(compressed = false, littleEndian = true, readHeaders = true)
+                ?: throw IOException("Could not load the NbtFile $file with any settings", ex)
+        }
     }
 }
 
@@ -101,112 +240,112 @@ private val NbtTag.typeId
     get() = serializers.indexOfFirst { it.kClass == this::class }
 
 private sealed class NbtSerial<T: NbtTag> (val kClass: KClass<T>){
-    abstract fun readTag(input: DataInputStream): T
-    abstract fun writeTag(output: DataOutputStream, tag: T)
+    abstract fun readTag(input: DataInput): T
+    abstract fun writeTag(output: DataOutput, tag: T)
 
     @Suppress("UNCHECKED_CAST")
     @JvmName("writeRawTag")
-    fun writeTag(output: DataOutputStream, tag: NbtTag) {
+    fun writeTag(output: DataOutput, tag: NbtTag) {
         writeTag(output, tag as T)
     }
 }
 
 private object NbtEndSerial: NbtSerial<NbtEnd>(NbtEnd::class) {
-    override fun readTag(input: DataInputStream): NbtEnd {
+    override fun readTag(input: DataInput): NbtEnd {
         return NbtEnd
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtEnd) {
+    override fun writeTag(output: DataOutput, tag: NbtEnd) {
     }
 }
 
 private object NbtByteSerial: NbtSerial<NbtByte>(NbtByte::class) {
-    override fun readTag(input: DataInputStream): NbtByte {
+    override fun readTag(input: DataInput): NbtByte {
         return NbtByte(input.readByte())
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtByte) {
+    override fun writeTag(output: DataOutput, tag: NbtByte) {
         output.writeByte(tag.signed.toInt())
     }
 }
 
 private object NbtShortSerial: NbtSerial<NbtShort>(NbtShort::class) {
-    override fun readTag(input: DataInputStream): NbtShort {
+    override fun readTag(input: DataInput): NbtShort {
         return NbtShort(input.readShort())
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtShort) {
+    override fun writeTag(output: DataOutput, tag: NbtShort) {
         output.writeShort(tag.value.toInt())
     }
 }
 
 private object NbtIntSerial: NbtSerial<NbtInt>(NbtInt::class) {
-    override fun readTag(input: DataInputStream): NbtInt {
+    override fun readTag(input: DataInput): NbtInt {
         return NbtInt(input.readInt())
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtInt) {
+    override fun writeTag(output: DataOutput, tag: NbtInt) {
         output.writeInt(tag.value)
     }
 }
 
 private object NbtLongSerial: NbtSerial<NbtLong>(NbtLong::class) {
-    override fun readTag(input: DataInputStream): NbtLong {
+    override fun readTag(input: DataInput): NbtLong {
         return NbtLong(input.readLong())
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtLong) {
+    override fun writeTag(output: DataOutput, tag: NbtLong) {
         output.writeLong(tag.value)
     }
 }
 
 private object NbtFloatSerial: NbtSerial<NbtFloat>(NbtFloat::class) {
-    override fun readTag(input: DataInputStream): NbtFloat {
+    override fun readTag(input: DataInput): NbtFloat {
         return NbtFloat(input.readFloat())
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtFloat) {
+    override fun writeTag(output: DataOutput, tag: NbtFloat) {
         output.writeFloat(tag.value)
     }
 }
 
 private object NbtDoubleSerial: NbtSerial<NbtDouble>(NbtDouble::class) {
-    override fun readTag(input: DataInputStream): NbtDouble {
+    override fun readTag(input: DataInput): NbtDouble {
         return NbtDouble(input.readDouble())
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtDouble) {
+    override fun writeTag(output: DataOutput, tag: NbtDouble) {
         output.writeDouble(tag.value)
     }
 }
 
 private object NbtByteArraySerial: NbtSerial<NbtByteArray>(NbtByteArray::class) {
-    override fun readTag(input: DataInputStream): NbtByteArray {
+    override fun readTag(input: DataInput): NbtByteArray {
         val size = input.readInt()
         val bytes = ByteArray(size)
         input.readFully(bytes)
         return NbtByteArray(bytes)
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtByteArray) {
+    override fun writeTag(output: DataOutput, tag: NbtByteArray) {
         output.writeInt(tag.value.size)
         output.write(tag.value)
     }
 }
 
 private object NbtStringSerial: NbtSerial<NbtString>(NbtString::class) {
-    override fun readTag(input: DataInputStream): NbtString {
+    override fun readTag(input: DataInput): NbtString {
         return NbtString(input.readUTF())
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtString) {
+    override fun writeTag(output: DataOutput, tag: NbtString) {
         output.writeUTF(tag.value)
     }
 }
 
 private object NbtListSerial: NbtSerial<NbtList<*>>(NbtList::class) {
-    override fun readTag(input: DataInputStream): NbtList<*> {
-        val type = input.read()
+    override fun readTag(input: DataInput): NbtList<*> {
+        val type = input.readUnsignedByte()
         val size = input.readInt()
         if (type == 0 && size > 0) {
             error("Missing type on NbtList")
@@ -219,7 +358,7 @@ private object NbtListSerial: NbtSerial<NbtList<*>>(NbtList::class) {
         return NbtList(list)
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtList<*>) {
+    override fun writeTag(output: DataOutput, tag: NbtList<*>) {
         val sample = tag.firstOrNull() ?: NbtEnd
         val typeId = sample.typeId
         val serializer = serializers[typeId]
@@ -237,10 +376,10 @@ private object NbtListSerial: NbtSerial<NbtList<*>>(NbtList::class) {
 }
 
 private object NbtCompoundSerial: NbtSerial<NbtCompound>(NbtCompound::class) {
-    override fun readTag(input: DataInputStream): NbtCompound {
+    override fun readTag(input: DataInput): NbtCompound {
         val map = mutableMapOf<String, NbtTag>()
         while (true) {
-            val typeId = input.read()
+            val typeId = input.readUnsignedByte()
             if (typeId == 0) {
                 break
             }
@@ -253,7 +392,7 @@ private object NbtCompoundSerial: NbtSerial<NbtCompound>(NbtCompound::class) {
         return NbtCompound(map)
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtCompound) {
+    override fun writeTag(output: DataOutput, tag: NbtCompound) {
         check(tag.values.none { it == NbtEnd }) {
             "NbtCompound cannot have an NbtEnd"
         }
@@ -271,7 +410,7 @@ private object NbtCompoundSerial: NbtSerial<NbtCompound>(NbtCompound::class) {
 }
 
 private object NbtIntArraySerial: NbtSerial<NbtIntArray>(NbtIntArray::class) {
-    override fun readTag(input: DataInputStream): NbtIntArray {
+    override fun readTag(input: DataInput): NbtIntArray {
         val size = input.readInt()
         val array = IntArray(size)
         for (i in 0 until size) {
@@ -280,7 +419,7 @@ private object NbtIntArraySerial: NbtSerial<NbtIntArray>(NbtIntArray::class) {
         return NbtIntArray(array)
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtIntArray) {
+    override fun writeTag(output: DataOutput, tag: NbtIntArray) {
         output.writeInt(tag.value.size)
         tag.value.forEach {
             output.writeInt(it)
@@ -289,7 +428,7 @@ private object NbtIntArraySerial: NbtSerial<NbtIntArray>(NbtIntArray::class) {
 }
 
 private object NbtLongArraySerial: NbtSerial<NbtLongArray>(NbtLongArray::class) {
-    override fun readTag(input: DataInputStream): NbtLongArray {
+    override fun readTag(input: DataInput): NbtLongArray {
         val size = input.readInt()
         val array = LongArray(size)
         for (i in 0 until size) {
@@ -298,7 +437,7 @@ private object NbtLongArraySerial: NbtSerial<NbtLongArray>(NbtLongArray::class) 
         return NbtLongArray(array)
     }
 
-    override fun writeTag(output: DataOutputStream, tag: NbtLongArray) {
+    override fun writeTag(output: DataOutput, tag: NbtLongArray) {
         output.writeInt(tag.value.size)
         tag.value.forEach {
             output.writeLong(it)